Update error handler
Errors are now helpful with stack traces in development mode, while they
are filtered out in deployment. This behavior, which we used to manually
implement, is now a thing by default.

CORS appears to not be handled by the Apollo handler, and instead gives
a response (and stack trace?) may need more investigation...